Changement dans la structure de la BD

+++ Create BD.sql
@@ -11,7 +11,7 @@ CREATE TABLE Etudiant (
 CREATE TABLE Semestre (
     idSemestre SMALLINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     intitule VARCHAR(255) NOT NULL CHECK(intitule IN ('semestre 1', 'semestre 2')),
-    annee TINYINT NOT NULL
+    annee SMALLINT(4) NOT NULL
 )ENGINE=INNODB;
 
 CREATE TABLE Cours (
